fix(raycaster): type facePoints as Float64Array
#205 changed the runtime raycast result to a Float64Array for full
precision but left the public type as Float32Array. Update the
RaycastResult type to match what is actually returned.

+++ packages/fragments/src/FragmentsModels/src/model/model-types.ts
@@ -352,7 +352,7 @@ export interface RaycastResult {
   /** The second edge of the snapped edge */
   snappedEdgeP2?: THREE.Vector3;
   /** The points of the raycasted face */
-  facePoints?: Float32Array;
+  facePoints?: Float64Array;
   /** The indices of the raycasted face */
   faceIndices?: Uint16Array;
 }
